feat(#60): 스터디 가입 신청 API 구현

+++ app/src/main/java/com/ssafy/neegongnaegong/data/datasource/network/NetworkStudiesDataSource.kt
@@ -21,7 +21,7 @@ interface NetworkStudiesDataSource {
 
     suspend fun deleteStudies(studyGroupId: Int): Flow<Unit>
 
-    suspend fun applyStudies(studyGroupId: Int): Flow<Unit>
+    suspend fun applyStudies(studyGroupId: Long): Flow<Unit>
 
     suspend fun cancelApplicationsStudies(studyGroupId: Int): Flow<Unit>
 }
